requirements.txt
Django>=2.1.3,<2.2.4
django-cors-headers==3.0.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djongo==1.2.32
gunicorn==19.9.0
PyJWT==1.7.1
pymongo==3.5.1
pytz==2019.1
sqlparse==0.2.4
whitenoise==4.1.2